+++ Assets/CommonCore/Core/DebugLog/DebugModule.cs
@@ -19,6 +19,11 @@ namespace CommonCore.DebugLog
             FPSCounter.Initialize();
         }
 
+        public override void Dispose()
+        {
+            FPSCounter.Terminate();
+        }
+
         public override void OnFrameUpdate()
         {
             if (!CCBase.Initialized)